+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/Ztm2025GeoMapViewer/OpenLayersGeoMap.jsx
@@ -3,7 +3,7 @@ import OLMap from 'ol/Map.js'
 import OpenStreetMaps from 'ol/source/OSM.js'
 import TileLayer from 'ol/layer/Tile.js'
 import View from 'ol/View.js'
-import OpenLayersCss from './OpenLayersCss.js'  // replaces import 'ol/ol.css'
+import OpenLayersCss from './OpenLayersCss.jsx'  // replaces import 'ol/ol.css'
 
 function OpenLayersGeoMap (props) {
   const olMapId = useId()
